norm(parent_proc): separate file
