+++ Makefile
@@ -15,6 +15,7 @@ EXEC_DIR	:=	exec
 SRCS		+=	$(EXEC_DIR)/check_command.c \
 				$(EXEC_DIR)/child_process.c \
 				$(EXEC_DIR)/exec.c \
+				$(EXEC_DIR)/parent_pipes.c \
 				$(EXEC_DIR)/parent_process.c
 
 INPUT_DIR	:=	input
