+++ includes/minishell.h
@@ -42,7 +42,9 @@ bool	is_first_command(int prev_fd);
 bool	is_last_command(char *next_cmd);
 void	child_process(t_command *cmd, t_fd *fd, char **environ);
 int		execute_command(t_command *commands);
-int		parent_process(t_command *cmd, t_fd *fd, pid_t pid, int *last_exit_status);
+int		handle_parent_pipes(t_command *cmd, t_fd *fd);
+int		parent_process(\
+					t_command *cmd, t_fd *fd, pid_t pid, int *last_exit_status);
 
 /* input */
 char	*input_line(void);
