+++ srcs/exec/parent_process.c
@@ -4,37 +4,8 @@
 #include "ft_dprintf.h"
 #include "libft.h"
 
-static int	handle_parent_pipes_except_first(t_fd *fd)
-{
-	if (x_close(fd->prev_fd) == PROCESS_ERROR)
-		return (PROCESS_ERROR);
-	return (EXIT_SUCCESS);
-}
-
-static int	handle_parent_pipes_except_last(t_fd *fd)
-{
-	fd->prev_fd = fd->pipefd[READ];
-	if (x_close(fd->pipefd[WRITE]) == PROCESS_ERROR)
-		return (PROCESS_ERROR);
-	return (EXIT_SUCCESS);
-}
-
-static int	handle_parent_pipes(t_command *cmd, t_fd *fd)
-{
-	if (!is_first_command(fd->prev_fd))
-	{
-		if (handle_parent_pipes_except_first(fd) == PROCESS_ERROR)
-			return (PROCESS_ERROR);
-	}
-	if (!is_last_command(*cmd->next_command))
-	{
-		if (handle_parent_pipes_except_last(fd) == PROCESS_ERROR)
-			return (PROCESS_ERROR);
-	}
-	return (EXIT_SUCCESS);
-}
-
-static int	get_last_command_status(pid_t pid, int *status, int *last_exit_status)
+static int	get_last_command_status(\
+								pid_t pid, int *status, int *last_exit_status)
 {
 	pid_t	wait_pid;
 
@@ -69,7 +40,8 @@ int	parent_process(t_command *cmd, t_fd *fd, pid_t pid, int *last_exit_status)
 
 	if (handle_parent_pipes(cmd, fd) == PROCESS_ERROR)
 		exit(EXIT_FAILURE);
-	if (get_last_command_status(pid, &status, last_exit_status) == PROCESS_ERROR)
+	if (get_last_command_status(pid, &status, last_exit_status) \
+															== PROCESS_ERROR)
 		return (PROCESS_ERROR);
 	if (wait_all_child_process(status) == PROCESS_ERROR)
 		return (PROCESS_ERROR);
